requirements.txt
dbus-python==1.3.2
mpris2==1.0.2
PyGObject==3.50.0
